Conflicts:
	SP4 Proposal and FrameWork/FrameWork/AI Assignment 1/AI Assignment
1.vcxproj.filters
	SP4 Proposal and FrameWork/FrameWork/AI Assignment 1/GamePlayState.h

+++ Personality/Personality/Personality.h
@@ -27,7 +27,7 @@ public:
 	}
 	float Vector3::Length(void) const
 	{
-		return sqrt(x * x + y * y );
+		return sqrt((double)x * x + y * y );
 	}
 };
